requirements.txt
boto3==1.28.46
botocore==1.31.46
pymongo==4.5.0
python-dotenv==1.0.1
